// ViaCloud/ClientApp/src/app/Models/Usuarios/usuarios.ts
export class Usuarios {
    usuarioID: number;
    tipoUsuarioID: number;
    riferoID: number;
    usuNombre: string;
    usuUserName: string;
    usuApellido: string;
    usuCedula: string;
    usuFechaNac: string;
    usuClave: string;
    usuComentario: string;
    usuNivel: number;
    usuActivo: boolean;
    nivelUsuarioPalabra: string;
    email: string;
    Tipo: string;


}